% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrying_capacity.R
\name{fit_cy_ds_models}
\alias{fit_cy_ds_models}
\title{Fit CY and DS Regression Models}
\usage{
fit_cy_ds_models(cyd, dsd)
}
\arguments{
\item{cyd}{A data frame with columns \code{Xri}, \code{Ygi}, and \code{Ydi}.}

\item{dsd}{A data frame with columns \code{Xej}, \code{Ygj}, and \code{Ydj}.}
}
\value{
A named list of linear models.
}
\description{
Fit CY and DS Regression Models
}
\examples{
ex <- r4c_example_data()
fit_cy_ds_models(ex$cyd, ex$dsd)
}
