% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot_validation}
\alias{plot_validation}
\title{Plot Validation Results}
\usage{
plot_validation(x, type = c("boxplot", "agreement"))
}
\arguments{
\item{x}{An object returned by \code{\link[=validate_methods]{validate_methods()}}.}

\item{type}{Plot type: \code{"boxplot"} or \code{"agreement"}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Validation Results
}
\examples{
ex <- r4c_example_data()
cc <- carrying_capacity(ex$cyd, ex$dsd)
v <- validate_methods(ex$cwd, cc)
plot_validation(v)
}
