% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4c-package.R
\docType{package}
\name{r4c-package}
\alias{r4c}
\alias{r4c-package}
\title{r4c: Range Condition and Carrying Capacity Tools}
\description{
User-friendly tools for:
}
\details{
\itemize{
\item range condition classification from SF data
\item carrying capacity estimation from CY and DS data
\item optional CW-based validation
}

The package accepts either Excel workbooks or ordinary \code{data.frame} objects.
}
\author{
\strong{Maintainer}: Mansour Mesdaghi \email{mmesdagh@yahoo.com}

}
\keyword{internal}
