% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_condition.R
\name{range_condition_class_rep}
\alias{range_condition_class_rep}
\title{Backward-Compatible Wrapper for Replicated Range Condition Input}
\usage{
range_condition_class_rep(sfd_or_list, PC, ...)
}
\arguments{
\item{sfd_or_list}{A single SF data frame or a list of replicated SF data
frames.}

\item{PC}{Palatability classes.}

\item{...}{Additional arguments passed to \code{\link[=classify_range_condition]{classify_range_condition()}}.}
}
\value{
An object of class \code{r4c_condition_result}.
}
\description{
Backward-Compatible Wrapper for Replicated Range Condition Input
}
