% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_cwd_excel}
\alias{read_cwd_excel}
\title{Read a CW Validation Sheet from an Excel Workbook}
\usage{
read_cwd_excel(path, sheet = "re_cwd")
}
\arguments{
\item{path}{Path to an Excel workbook.}

\item{sheet}{Sheet name. Defaults to \code{"re_cwd"}.}
}
\value{
A data frame.
}
\description{
Read a CW Validation Sheet from an Excel Workbook
}
