% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_cyd_excel}
\alias{read_cyd_excel}
\alias{read_ccd_excel}
\title{Read a CY Sheet from an Excel Workbook}
\usage{
read_cyd_excel(path, sheet = "cyd")

read_ccd_excel(path, sheet = "cyd")
}
\arguments{
\item{path}{Path to an Excel workbook.}

\item{sheet}{Sheet name. Defaults to \code{"cyd"}.}
}
\value{
A data frame.
}
\description{
Read a CY Sheet from an Excel Workbook
}
