% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_dsd_excel}
\alias{read_dsd_excel}
\alias{read_dpd_excel}
\title{Read a DS Sheet from an Excel Workbook}
\usage{
read_dsd_excel(path, sheet = "dsd")

read_dpd_excel(path, sheet = "dsd")
}
\arguments{
\item{path}{Path to an Excel workbook.}

\item{sheet}{Sheet name. Defaults to \code{"dsd"}.}
}
\value{
A data frame.
}
\description{
Read a DS Sheet from an Excel Workbook
}
