% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_methods}
\alias{validate_methods}
\title{Validate CY and DS Estimates Against CW Data}
\usage{
validate_methods(
  cw_df,
  models,
  test = c("anova", "kruskal"),
  cw_sheet = "re_cwd"
)
}
\arguments{
\item{cw_df}{A CW validation data frame, an Excel workbook path, or \code{NULL}.}

\item{models}{Either the output of \code{\link[=fit_cy_ds_models]{fit_cy_ds_models()}} or
\code{\link[=carrying_capacity]{carrying_capacity()}}.}

\item{test}{Statistical test to apply.}

\item{cw_sheet}{Sheet name used when \code{cw_df} is a workbook path.}
}
\value{
An object of class \code{r4c_validation_result}.
}
\description{
Uses CW data to compare corrected dry matter estimates from the CY and DS
models. Validation can be done using ANOVA or Kruskal-Wallis.
}
\examples{
ex <- r4c_example_data()
cc <- carrying_capacity(ex$cyd, ex$dsd)
validate_methods(ex$cwd, cc)
}
